+++ modules/apim-adapter/src/test/resources/com/axway/apim/model/BEAPICreatedWithCreatedOn.json
