+++ modules/apim-adapter/src/test/resources/com/axway/apim/model/ConfiguredOutboundProfile1.json
