+++ modules/apim-adapter/src/test/resources/com/axway/apim/model/ConfiguredOutboundProfile2.json
